Basic Layout for "Big Llama Tours" website
Set basic page layout using Flexbox

Add basic text and images for guidance and layout purposes. Intended to be changed later

This commit will be the basic version of the webpage. The design was provided by TOP, any improvement or changes from
now on will be based on this design